+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.2] - 2026-09-23
+
 ### Changed
 
 - Decoding data records needs less stack and time. On a Cortex-M4, peak stack